Update code.py
The new changes has been done by Kalpak Kanungo. Now we can run this program multiple times by the user saying "YES" or "No". The user can run this program as many times as he wants so if the user says "YES" the program will simply run from its initial position and if "NO" the program will simply terminate and the mail will be send to the receiver. The user interface has been made friendly and easy to understand for the developer.